Fixing problem where a 0-length list will cause a crash in CASUALcore when saving from CASCADE

+++ CASUALcore/src/CASUAL/caspac/Caspac.java
@@ -214,21 +214,23 @@ public final class Caspac {
         }
         for (Script s : scripts) {
             //individualFiles.toArray();
-            for (File test: new File(s.tempDir).listFiles()){
-                boolean delete=true;
-                for (File f:s.individualFiles){
-                    if (test.getCanonicalFile().equals(f.getCanonicalFile())){
-                        delete=false;
-                    }
-                }    
-                if (delete){
-                    if (test.toString().contains(s.tempDir)){
-                        test.delete();
-                    }
-                    
-                } 
+            File[] list=new File(s.tempDir).listFiles();
+            if (list!=null){
+                for (File test: list){
+                    boolean delete=true;
+                    for (File f:s.individualFiles){
+                        if (test.getCanonicalFile().equals(f.getCanonicalFile())){
+                            delete=false;
+                        }
+                    }    
+                    if (delete){
+                        if (test.toString().contains(s.tempDir)){
+                            test.delete();
+                        }
+
+                    } 
+                }
             }
-            
             nameStream.putAll(s.getScriptAsMapForCASPAC());
         }
